requirements.txt
Django
django-oauth-toolkit 
djangorestframework